Additional questions part works now; Search related stuff are in same package now.

+++ domain/src/main/java/com/koonat/ouluhealth/domain/model/Diagnosis.kt
@@ -19,7 +19,7 @@ data class PredictionBody(
 data class PredictedSymptom(
         @SerializedName("id") val id: Int,
         @SerializedName("name") val name: String,
-        @SerializedName("score") var score: Int,
+        @SerializedName("score") var score: Double,
         @SerializedName("positive") var positive: Boolean = false,
         @SerializedName("skip") var skip: Boolean = false
 )
